Move tests to outside of the production code
It's standard in python

Bug: T219360
